--- a/test/data/Makefile
+++ b/test/data/Makefile
@@ -8,6 +8,7 @@ OSRM_EXTRACT:=$(TOOL_ROOT)/osrm-extract
 OSRM_CONTRACT:=$(TOOL_ROOT)/osrm-contract
 OSRM_ROUTED:=$(TOOL_ROOT)/osrm-routed
 POLY2REQ:=$(SCRIPT_ROOT)/poly2req.js
+MD5SUM:=$(SCRIPT_ROOT)/md5sum.js
 TIMER:=$(SCRIPT_ROOT)/timer.sh
 PROFILE:=$(PROFILE_ROOT)/car.lua
 
@@ -24,7 +25,7 @@ $(DATA_NAME).poly:
 
 $(DATA_NAME).osrm: $(DATA_NAME).osm.pbf $(DATA_NAME).poly $(PROFILE) $(OSRM_EXTRACT)
 	@echo "Verifiyng data file integrity..."
-	md5sum -c data.md5sum
+	$(MD5SUM) -c data.md5sum
 	@echo "Running osrm-extract..."
 	$(TIMER) "osrm-extract" $(OSRM_EXTRACT) $(DATA_NAME).osm.pbf -p $(PROFILE)
 
@@ -49,6 +50,6 @@ benchmark: $(DATA_NAME).requests osrm-routed.pid
 	@echo "****************"
 
 checksum:
-	md5sum $(DATA_NAME).osm.pbf $(DATA_NAME).poly > data.md5sum
+	$(MD5SUM) $(DATA_NAME).osm.pbf $(DATA_NAME).poly > data.md5sum
 
 .PHONY: clean checksum benchmark