As we know, OS X builds on travis are slow. Today [a build took 30 minutes](https://travis-ci.org/Project-OSRM/osrm-backend/jobs/174879851). For this build the longest running steps, sorted by time, are below:

 - 754s (npm test)
 - 235s (brew install)
 - 104s (make osrm-extract --jobs=3)
 - 98s (make --jobs=3)
 - 57s (npm install)
 - 47s (./unit_tests/util-tests)
 - 32s (make tests)
 - 32s (make install)

Of those the longest `npm test` is easy to improve upon. We can build in `Release` mode instead of `Debug` to ensure the tests run much faster. This PR therefore moves the `osx` build to `Release` mode.

Also, of those the second longest is also easy to improve upon. Instead of using `brew` we can use mason deps pulled from s3. This PR therefore moves the `osx` build to use mason.

## Tasklist
 - [x] review
 - [x] adjust for comments