@springmeyer 2016-11-11
Currently hitting:

```
undef: _clock_gettime
Undefined symbols for architecture x86_64:
  "_clock_gettime", referenced from:
      boost::interprocess::spin_wait::yield() in lto.o
ld: symbol(s) not found for architecture x86_64
clang-3.8: error: linker command failed with exit code 1 (use -v to see invocation)
```