@danpat 2016-11-11
@springmeyer ugh - could this be related to our OSX min version of 10.10 ?

Looking at SO: http://stackoverflow.com/questions/5167269/clock-gettime-alternative-in-mac-os-x
it seems like clock_gettime re-appeared in OSX 10.12.  What version of OSX SDK were the mason binaries built against?  It might be that if they were built in a 10.12 environment, some conditionals inside boost have included a symbol that we don't want on older platforms.