@springmeyer 2016-11-11
@danpat - What I'm seeing is that the symbol problem is in `boost::interprocess`. This `mason` package is header only so there are no binaries involved. I don't know if this is related to our "min version". My guess is that it is a boost interprocess header-detection bug and that
https://github.com/boostorg/interprocess/blob/master/include/boost/interprocess/detail/os_thread_functions.hpp#L343 might need patched for 10.12.