@springmeyer 2016-11-11
hmm, or we need to set `-no_weak_imports`. Will plan to digest and test https://github.com/Homebrew/homebrew-core/issues/3727 more next week.