@springmeyer 2016-11-11
This branch is now working by using `xcode7.3`. Will rebase shortly. Ticketed the `clock_gettime` problem separately: #3297