@springmeyer 2016-11-12
> build took 30 minutes

This PR is now building in less than half the time on OSX: 12 minutes. Here are the new timings:

 - 754s (npm test) -> 371s
 - 235s (brew install) -> 32s (to install mason deps during `cmake ../`)
 - 104s (make osrm-extract --jobs=3) ->42s
 - 98s (make --jobs=3) -> 43s
 - 57s (npm install) -> 23s
 - 47s (./unit_tests/util-tests) ->  16s
 - 32s (make tests) -> 43s
 - 32s (make install) -> 6s