@springmeyer 2016-11-12
This is ready for review. A summary is:

  - Now building in `Release` on OSX instead of `Debug` for faster test runs
  - Now avoiding using `brew` to install dependencies to save ~4min
    - This exposed a few things that could be improved for both linux and osx:
      - Putting `ccache` from mason on PATH to avoid needing to install globally (since travis calls ccache directly)
      - Dropping the dependency on md5sum like we did in node-osrm: https://github.com/Project-OSRM/node-osrm/pull/258
      - Since we now vendor mason, this PR moves to using the mason client rather than the custom access to mason packages for `ccache` and `cmake`