@TheMarex 2016-11-12
While running in `Debug` is always from advantage because of assertions, I can see that the OSX situation is just unmanageable. Hopefully we can revert back to `Debug` one day when travis is a little faster. Otherwise :+1: