@springmeyer 2016-11-12
> Hopefully we can revert back to Debug one day when travis is a little faster

👍  Or should we consider adding a `Release+Assertions` mode?