@TheMarex 2016-11-13
Yeah good idea. Would also help to actually test that option on travis. :+1: