@springmeyer 2016-11-14
> Yeah good idea. Would also help to actually test that option on travis. 👍

Queued up PR in https://github.com/Project-OSRM/osrm-backend/pull/3305.